backend/login_server/requirements.txt
Flask==3.0.3
# Flask-OAuthlib 
# Flask-RESTful 
# Flask-JWT-Extended 
pymongo
authlib==1.3.1
google-auth-oauthlib
google-auth
# mongoengine
pyjwt
pymongo[srv]
python-dotenv
flask_cors
bcrypt
PyJWT